refactor: explode model do backend em suas respectivas features
