+++ src/features/acf/backend/diabetes/modules/externalCards/adapter.ts
@@ -3,7 +3,7 @@ import type {
     AcfExternalCardsDescription,
     ExternalCardDataItem,
     ExternalCardDBDataItem,
-} from "../../model";
+} from "./model";
 
 export const externalCardsDbToModel = (
     data: Array<ExternalCardDBDataItem>
