+++ src/features/acf/backend/diabetes/modules/externalCards/model.ts
@@ -1,0 +1,19 @@
+import type { AcfDashboardType } from "@/features/acf/shared/diabetes/model";
+
+export type AcfExternalCardsDescription =
+    | "DIAGNOSTICO_AUTORREFERIDO"
+    | "DIAGNOSTICO_CLINICO";
+
+export type ExternalCardDataItem = {
+    acfDashboardType: AcfDashboardType;
+    acfExternalCardsDescription: AcfExternalCardsDescription;
+    value: number;
+};
+
+export type ExternalCardDBDataItem = {
+    municipio_id_sus: string;
+    ine: string;
+    lista: string;
+    valor: number;
+    descricao: string;
+};
