+++ src/features/acf/backend/diabetes/modules/table/repository.ts
@@ -1,7 +1,7 @@
 import type { SortOrder } from "@/features/acf/shared/diabetes/model";
 import type { impulso_previne_dados_nominais___painel_enfermeiras_lista_nominal_diabeticos } from "@prisma/client";
 import { prisma } from "@prisma/prismaClient";
-import type { DiabetesDbFilterItem, SortableDbField } from "../../model";
+import type { DiabetesDbFilterItem, SortableDbField } from "./model";
 const pageSize = 8;
 
 type QueryWhere = {
